Refactor initial CLI
