+++ .eslintrc.js
@@ -12,5 +12,7 @@ module.exports = {
     "prettier",
     "plugin:prettier/recommended",
   ],
-  rules: {},
+  rules: {
+    "no-console": "off",
+  },
 };
